урок 7 | использовали okhttp для синхронного(!) запроса

+++ app/src/main/java/com/gb/k_2135_2136_2/model/dto/Part.kt
@@ -22,7 +22,7 @@ data class Part(
     @SerializedName("polar")
     val polar: Boolean,
     @SerializedName("prec_mm")
-    val precMm: Int,
+    val precMm: Double,
     @SerializedName("prec_period")
     val precPeriod: Int,
     @SerializedName("prec_prob")
